Dedupe overlay status SSE handler into shared sseChannel helper
qlty flagged the /overlay/settings/events and /alerts/settings/events
handlers as near-duplicate (36 shared lines). Factored the streamer
resolution + poll-and-broadcast-on-change lifecycle into
createOverlayStatusEventsHandler in sseChannel.ts, mirroring the existing
createStreamerSseEventsHandler pattern.
