+++ src/web/routes/alertsAdmin.ts
@@ -4,14 +4,13 @@ import { csrfProtection } from '../csrf';
 import { requireAuth } from '../middleware';
 import { getSessionUser } from '../session';
 import { getStreamerByDiscordId, getAlertConfigsForStreamer, ALERT_EVENT_TYPES, ALERT_TEXT_ANIMATIONS } from '../../db';
-import type { DbStreamerEventSub } from '../../db';
 import { PUBLIC_URL, ALERT_STATUS_MAX_SSE_PER_STREAMER } from '../../shared/config';
 import { filterQueryParam } from './validation';
 import { renderError, renderView } from './viewHelpers';
 import { router as mutationsRouter } from './alertsAdminMutations';
 import { router as assetMutationsRouter, MAX_IMAGE_MB, MAX_SOUND_MB } from './alertsAssetMutations';
 import { connections as alertsSourceConnections } from './alertsOverlaySource';
-import { attachSseConnection, broadcastToChannel } from './sseChannel';
+import { createOverlayStatusEventsHandler } from './sseChannel';
 
 const log = createLogger('AlertsAdmin');
 const router = Router();
@@ -67,51 +66,22 @@ router.get('/settings', requireAuth, csrfProtection, async (req, res) => {
  * GET /alerts/settings/events — SSE endpoint streaming `{ connected: boolean }` snapshots of
  * whether the logged-in user's own alerts browser source currently has an open connection, so the
  * settings page can show a live status dot instead of the user only finding out something's wrong
- * when an alert never fires. Polls the shared `alertsOverlaySource` connections map on an interval
- * rather than reacting to a push event, since opening/closing that overlay's SSE connection has no
- * existing event to subscribe to.
+ * when an alert never fires. Connection lifecycle and status-polling logic (streamer resolution,
+ * SSE handshake, poll-and-broadcast-on-change, and interval cleanup) is shared with the
+ * reward-video overlay via `createOverlayStatusEventsHandler`.
  * @param req - Express request; reads `req.session.user`.
  * @param res - Express response; upgrades to a `text/event-stream` connection kept alive with
  *   periodic pings and torn down (including the status-poll interval) on client disconnect;
  *   replies 403 if the user isn't a monitored streamer with a linked Twitch channel, 500 (logged)
  *   if the streamer lookup fails, or 429 if the connection limit is exceeded.
  */
-router.get('/settings/events', requireAuth, async (req, res) => {
-  let streamer: DbStreamerEventSub | null;
-  try {
-    streamer = await getStreamerByDiscordId(getSessionUser(req).discordId);
-  } catch (err) {
-    log.error('Failed to resolve streamer for alerts status SSE:', err);
-    res.status(500).end();
-    return;
-  }
-  if (!streamer || !streamer.twitch_name) {
-    res.status(403).end();
-    return;
-  }
-
-  const attached = attachSseConnection(req, res, {
-    connections: statusConnections,
-    key: streamer.id,
-    maxPerChannel: ALERT_STATUS_MAX_SSE_PER_STREAMER,
-  });
-  if (!attached) return;
-
-  const streamerId = streamer.id;
-  const login = streamer.twitch_name.toLowerCase();
-  let lastConnected: boolean | null = null;
-
-  const check = (): void => {
-    const isConnected = (alertsSourceConnections.get(login)?.size ?? 0) > 0;
-    if (isConnected === lastConnected) return;
-    lastConnected = isConnected;
-    broadcastToChannel(statusConnections, streamerId, { connected: isConnected });
-  };
-
-  check();
-  const interval = setInterval(check, STATUS_POLL_INTERVAL_MS);
-  req.on('close', () => clearInterval(interval));
-});
+router.get('/settings/events', requireAuth, createOverlayStatusEventsHandler({
+  statusConnections,
+  overlayConnections: alertsSourceConnections,
+  maxPerChannel: ALERT_STATUS_MAX_SSE_PER_STREAMER,
+  pollIntervalMs: STATUS_POLL_INTERVAL_MS,
+  log,
+}));
 
 router.use(mutationsRouter);
 router.use(assetMutationsRouter);
